Fix all CI failures: ruff, formatting, and mypy strict compliance
- Remove black from CI (conflicts with ruff format); use ruff as canonical formatter
- Fix ruff check errors in experiments/ (E402 noqa for sys.path imports)
- Fix ruff format across all source and test files
- Fix all 112 mypy --strict errors:
  - Add type annotations for numpy return values (no-any-return)
  - Fix variable type conflicts in if/else branches (assignment)
  - Add base class type annotations for polymorphic variables
  - Convert str to ClientId and DemographicDistribution to NDArray at boundaries
  - Add get_config() to NoiseMechanism ABC
  - Declare Optional types for nullable instance attributes
- All 503 tests pass, 0 ruff errors, 0 mypy errors

+++ src/fairswarm/algorithms/result.py
@@ -72,10 +72,8 @@ class ConvergenceMetrics:
         if len(recent) < 2:
             return 0.0
 
-        improvements = [
-            recent[i] - recent[i - 1] for i in range(1, len(recent))
-        ]
-        return np.mean(improvements)
+        improvements = [recent[i] - recent[i - 1] for i in range(1, len(recent))]
+        return float(np.mean(improvements))
 
 
 @dataclass
@@ -118,10 +116,7 @@ class FairnessMetrics:
         if not self.target_distribution:
             return 0.0
 
-        gaps = [
-            self.representation_gap(group)
-            for group in self.target_distribution
-        ]
+        gaps = [self.representation_gap(group) for group in self.target_distribution]
         return max(gaps) if gaps else 0.0
